Fix bugs discovered during testing PR #1: update operator allowance in the functions and not in the modifiers (#51)
@Stebalien This causes all sorts of issues.

For example, when a rail in settled, the account lockup changes but that
changes should not reflect in the operator lockup usage as lockup usage
is meant to track "future" lockup.

But we still want to use the invariant modifier during rail settlement
as well.

Overall, mutating state in modifers is an anti-pattern and I see why and
so want to keep it at an absolute minimum.