Fix bugs discovered during testing PR #2: Allow Rail settlement and foreclosure (#52)
@Stebalien 

- I don't think we should allow people to withdraw the rail lockup
during normal settlement. It messes up all of our accounting. When you
settle a rail in the normal cause of events, you can only withdraw upto
and including the account lockup.

- If you want to withdraw the rail "future" lockup as well (might not be
the "future" for a rail in debt) -> you should foreclose the rail and be
done with it as you are essentially saying that you don't trust the
client to deposit more funds anyways.

---------

Co-authored-by: Steven Allen <steven@stebalien.com>