feat: add support for native FIL (#73)
This adds support for native FIL tokens by special-casing 0x0 on both
deposit and withdraw.

This is the simplest possible approach. We might want to make it nicer
by introducing separate withdraw/deposit methods but... I'm not actually
sure if that's necessary or even nicer.

NOTE: 0x0 is the address of the Ethereum system actor and no contract
will ever be deployed there, so this is always safe.

fixes #72

TODO:

- [x] write tests