Optimize the rate-change queue (#86)
This stores the rate-change queue as a single dynamic array instead of a
mapping. Given how both the EVM and FEVM work, there's no extra cost for
sparse arrays but there is for mappings (especially in FEVM).

From EVM gas benchmarking, this doesn't appear to make much of a
difference (and I'm really not clear on why) but it makes the queue take
up fewer slots at the very least.

In the FVM, this should ensure that rate changes are stored adjacent to
eachother making settlement less expensive.