fix: use 'to' parameter instead of msg.sender in depositWithPermit (#142)
This change makes it compatible with USDFC.

I have to admit to not having a super solid theory as to why this works
and using `msg.sender` doesn't, but when I switch it out I can make it
work, but with `msg.sender` I get a [`DebtToken: invalid
signature`](https://github.com/Secured-Finance/stablecoin-contracts/blob/develop/contracts/DebtToken.sol#L182-L211)
failure.

This change seems to fix the issue by using the `to` parameter instead
of `msg.sender` in the permit call. Since `to` is already the owner who
signed the permit, this is semantically correct. BUT since they should
be the same value when we make this I don't have a good explanation as
to why this makes a difference, but it does.

Deploying a contract with this change on calibnet lets me do a permit
sign + single transaction to get setup with USDFC.

@pali101 perhaps you have some insight or ideas here?

---------

Co-authored-by: aarshkshah1992 <aarshkshah1992@gmail.com>