fix: don't unlock client funds for the full requested settlement duration if arbitrer reduces settled duration (#157)
https://github.com/FilOzone/filecoin-services-payments/pull/141 fixed
the bug where client funds do not get unlocked if arbitrer settles the
full duration but not the full amount i.e. reduced the requested
settlement amount.

This PR fixes the other bug where arbitrer can reduce the settled
duration as well. In this case, we should only unlock client funds for
the duration settled by the arbitrer and not for the entire requested
settlement duration.