Audit: make deposit accounting safe (#165)
Currently deposit accounting trusts that the ERC20 will transfer the
entire amount from account to account. While that is best practice and
probably holds for usfdc it is not a guarantee and so before and after
balance accounting is needed to prevent the payments contract from
getting into a bad state where it thinks it has more tokens than it
actually does.

---------

Co-authored-by: zenground0 <ZenGround0@users.noreply.github.com>