Fixup README (#163)
A few small README fixups in preparation for #148

### Rename arbiter to validator 

Note I kept a small number of references to validation as
validation/arbitration. Eva wanted the rename because in the common case
there is no arbitration. But in the generic case arbitration is a good
description of what is happening and theres no need to scrub the word
from the README entirely.

### Add functions to interface 
* getAccountInfoIfSettled

---------

Co-authored-by: zenground0 <ZenGround0@users.noreply.github.com>